+++ src/pynxtools/eln_mapper/reader_eln.py
@@ -20,8 +20,8 @@ is generated by schema eln."""
 
 import re
 
-from pynxtools.dataconverter.nexus_tree import NexusEntity, NexusGroup, NexusNode
 from pynxtools.eln_mapper.eln import ElnGenerator
+from pynxtools.nexus.nexus_tree import NexusEntity, NexusGroup, NexusNode
 
 
 class ReaderElnGenerator(ElnGenerator):
