+++ src/pynxtools/eln_mapper/schema_eln.py
@@ -20,8 +20,8 @@
 import re
 from typing import Union
 
-from pynxtools.dataconverter.nexus_tree import NexusEntity, NexusGroup, NexusNode
 from pynxtools.eln_mapper.eln import ElnGenerator
+from pynxtools.nexus.nexus_tree import NexusEntity, NexusGroup, NexusNode
 
 NEXUS_TO_NOMAD_QUANTITY: dict[str, tuple[str, str]] = {
     "NX_BINARY": ("bytes", "NumberEditQuantity"),
